refactor: 경로 조회 타입으로 PathType을 반환하는 기능 메서드명 변경
